+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog] and this project adheres to
 [Semantic Versioning], with the exception that v10 is version 0 in
 semver parlance.
 
+[Unreleased]
+------------
+
+### Added
+
+-   `require`, with RUBSH_PATH
+
 [v10.10.11] - 2017-04-07
 ------------------------
 
@@ -86,5 +93,6 @@ semver parlance.
 
   [Keep a Changelog]: http://keepachangelog.com/
   [Semantic Versioning]: http://semver.org/
+  [Unreleased]: https://github.com/binaryphile/rubsh/compare/v10.10.11...v10.11
   [v10.10.11]: https://github.com/binaryphile/rubsh/compare/v10.10.10...v10.10.11
   [v10.10.10]: https://github.com/binaryphile/rubsh/compare/v0.6.3...v10.10.10